restructured rdf_main_generator.rb
