+++ lib/jekyll-rdf.rb
@@ -44,6 +44,7 @@ require 'jekyll/filters/rdf_sparql_query'
 require 'jekyll/filters/rdf_property'
 require 'jekyll/filters/rdf_collection'
 require 'jekyll/filters/rdf_container'
+require 'jekyll/rdf_generator_helper'
 require 'jekyll/rdf_main_generator'
 require 'jekyll/rdf_page_helper'
 require 'jekyll/rdf_page_data'
